refactor(uninstall): use dynamic import for subcommand
