+++ packages/opencode/src/index.ts
@@ -7,7 +7,7 @@ import { ConsoleCommand } from "./cli/cmd/account/command"
 import { ProvidersCommand } from "./cli/cmd/providers/command"
 import { AgentCommand } from "./cli/cmd/agent/command"
 import { UpgradeCommand } from "./cli/cmd/upgrade"
-import { UninstallCommand } from "./cli/cmd/uninstall"
+import { UninstallCommand } from "./cli/cmd/uninstall/command"
 import { ModelsCommand } from "./cli/cmd/models/command"
 import { UI } from "./cli/ui"
 import { Installation } from "./installation"
